# docs/swagger.yaml
basePath: /
definitions:
  handlers.HealthResponse:
    properties:
      status:
        type: string
    type: object
  models.Product:
    properties:
      id:
        type: integer
      name:
        minLength: 2
        type: string
      price:
        type: number
    required:
    - name
    - price
    type: object
host: localhost:8080
info:
  contact: {}
  description: A simple Golang REST application featuring products and health check.
  title: Golang REST API
  version: "1.0"
paths:
  /health:
    get:
      consumes:
      - application/json
      description: get the status of server
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HealthResponse'
      summary: Show the status of server
      tags:
      - health
  /products:
    get:
      consumes:
      - application/json
      description: get products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: add by json product
      parameters:
      - description: Add product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Add a product
      tags:
      - products
  /products/batch:
    post:
      consumes:
      - application/json
      description: Accepts an array of products and queues them for background processing
      parameters:
      - description: Array of products
        in: body
        name: products
        required: true
        schema:
          items:
            $ref: '#/definitions/models.Product'
          type: array
      produces:
      - application/json
      responses:
        "202":
          description: Accepted for processing
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Add multiple products in background
      tags:
      - products
  /ready:
    get:
      consumes:
      - application/json
      description: pings the database to verify readiness
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HealthResponse'
        "503":
          description: Service Unavailable
          schema:
            type: string
      summary: Show the readiness of server
      tags:
      - health
swagger: "2.0"